+++ src/compiler.rs
@@ -1,8 +1,10 @@
 //! Implementation of the compiler inside the `Jsonpiler`.
 use super::{
-  Align, Args, AsmFunc, Bind, Builtin, ErrOR, FuncInfo, JFunc, JObject, JResult, Json, JsonWithPos,
-  Jsonpiler, Position, Section, err,
-  utility::{fmt_local, scope_begin, scope_end},
+  Args, AsmFunc, Bind, Builtin, ErrOR, FuncInfo, GVar, JFunc, JObject, JResult, Json, JsonWithPos,
+  Jsonpiler, Name, Position,
+  Var::{Global, Tmp},
+  add, err,
+  utility::{imp_call, mn, scope_begin, scope_end},
 };
 use core::mem;
 use std::{
@@ -25,24 +27,26 @@ impl Jsonpiler {
   /// # Errors
   /// * `Box<dyn Error>` - If an error occurred during the compilation.
   #[inline]
-  pub fn build(&mut self, source: String, json_file: &str, out_file: &str) -> ErrOR<()> {
+  pub fn build(&mut self, source: String, out_file: &str) -> ErrOR<()> {
     let json = self.parse(source)?;
     self.include_flag = HashSet::new();
-    self.sect = Section::default();
-    self.symbol_seeds = HashMap::new();
+    self.text = vec![];
+    self.bss = vec![];
+    self.data = vec![];
+    self.global_seed = 0;
     self.builtin = HashMap::new();
     self.vars = vec![HashMap::new()];
     self.all_register();
     let mut info = FuncInfo::default();
     let result = self.eval(json, &mut info)?;
     let mut writer = io::BufWriter::new(File::create(out_file)?);
-    writeln!(writer, ".file \"{json_file}\"\n.intel_syntax noprefix")?;
+    writer.write_all(mn(".intel_syntax", &["noprefix"]).as_bytes())?;
     writer.write_all(include_bytes!("asm/once/data.s"))?;
-    for data in &mut self.sect.data {
+    for data in &mut self.data {
       writer.write_all(data.as_bytes())?;
     }
     writer.write_all(include_bytes!("asm/once/bss.s"))?;
-    for bss in &mut self.sect.bss {
+    for bss in &mut self.bss {
       writer.write_all(bss.as_bytes())?;
     }
     writer.write_all(include_bytes!("asm/once/main.s"))?;
@@ -53,17 +57,17 @@ impl Jsonpiler {
     }
     if let Json::Int(int) = result.value {
       match int {
-        Bind::Lit(l_int) => writeln!(writer, "  mov rcx, {l_int}"),
-        Bind::Var(var) => writeln!(writer, "  mov rcx, {var}"),
-        Bind::Local(local) | Bind::Tmp(local) => {
-          writeln!(writer, "  mov rcx, {}", fmt_local("qword", local))
+        Bind::Lit(l_int) => writer.write_all(mn("mov", &["rcx", &l_int.to_string()]).as_bytes()),
+        Bind::Var(name) => {
+          writer.write_all(mn("mov", &["rcx", &format!("qword{name}")]).as_bytes())
         }
       }
     } else {
-      writeln!(writer, "  xor ecx, ecx")
+      writer.write_all(mn("xor", &["ecx", "ecx"]).as_bytes())
     }?;
-    writer.write_all(b"  call [qword ptr __imp_ExitProcess[rip]]\n.seh_endproc\n")?;
-    for text in &mut self.sect.text {
+    writer.write_all(imp_call("ExitProcess").as_bytes())?;
+    writer.write_all(mn(".seh_endproc", &[]).as_bytes())?;
+    for text in &mut self.text {
       writer.write_all(text.as_bytes())?;
     }
     writer.write_all(include_bytes!("asm/once/handler.s"))?;
@@ -82,8 +86,8 @@ impl Jsonpiler {
       return Ok(JsonWithPos { value: Json::Object(Bind::Lit(evaluated)), pos: json.pos });
     };
     self.validate_args("function call", true, 1, list.len(), &json.pos)?;
-    let first_elem = mem::take(list.first_mut().ok_or(ERR)?);
-    let first = self.eval(first_elem.clone(), info)?;
+    let raw_first = mem::take(list.first_mut().ok_or(ERR)?);
+    let first = self.eval(raw_first.clone(), info)?;
     if let Json::String(Bind::Lit(cmd)) = &first.value {
       if self.builtin.contains_key(cmd) {
         let builtin = self.builtin.get_mut(cmd).ok_or(ERR)?;
@@ -103,50 +107,43 @@ impl Jsonpiler {
           scope_end(&mut tmp, info)?;
           self.vars.pop();
         }
-        Ok(JsonWithPos { value: result, pos: first_elem.pos })
+        Ok(JsonWithPos { value: result, pos: json.pos })
       } else if let Ok(Json::Function(af)) = self.get_var(cmd, &first.pos) {
-        call_func(first_elem.pos, &af, info)
+        call_func(json.pos, &af, info)
       } else {
         err!(self, first.pos, "The `{cmd}` function is undefined.")
       }
     } else if let Json::Function(af) = &first.value {
-      call_func(first_elem.pos, af, info)
+      call_func(json.pos, af, info)
     } else {
-      self.typ_err(1, "function call", "LString` or `Function", &first_elem)?;
+      self.typ_err(1, "function call", "LString` or `Function", &raw_first)?;
       Ok(JsonWithPos::default())
     }
   }
   /// Evaluate arguments.
-  fn eval_args(&mut self, args: Args, info: &mut FuncInfo) -> ErrOR<Vec<JsonWithPos>> {
-    let mut result = vec![];
-    for arg in args {
-      result.push(self.eval(arg, info)?);
+  fn eval_args(&mut self, mut args: Args, info: &mut FuncInfo) -> ErrOR<Vec<JsonWithPos>> {
+    for arg in &mut args {
+      *arg = self.eval(mem::take(arg), info)?;
     }
-    Ok(result)
+    Ok(args)
   }
   /// Generates a unique name for internal use.
-  pub(crate) fn get_global(&mut self, name: &str, value: &str) -> ErrOR<String> {
-    let seed = self
-      .symbol_seeds
-      .get(name)
-      .map_or(Ok(0), |current| current.checked_add(1).ok_or("Seed Overflow"))?;
-    self.symbol_seeds.insert(name.to_owned(), seed);
-    let l_name = format!(".L{name}{seed:x}");
-    let name_fmt = format!("qword ptr {l_name}[rip]");
+  pub(crate) fn get_global(&mut self, name: &GVar, value: &str) -> ErrOR<Name> {
+    let seed = self.global_seed;
     match name {
-      "BSS" => self.sect.bss.push(format!("  .lcomm {l_name}, {value}\n")),
-      "STR" => {
+      GVar::Bss => self.bss.push(format!("  .lcomm .L{seed:x}, {value}\n")),
+      GVar::Str => {
         if let Some(str_seed) = self.str_cache.get(value) {
-          return Ok(format!("qword ptr .LSTR{str_seed:x}[rip]"));
+          return Ok(Name { var: Global, seed: *str_seed });
         }
-        self.str_cache.insert(value.to_owned(), self.symbol_seeds.len());
-        self.sect.data.push(format!("  {l_name}: .string \"{value}\"\n"));
+        self.str_cache.insert(value.to_owned(), seed);
+        self.data.push(format!("  .L{seed:x}: .string \"{value}\"\n"));
       }
-      "INT" => self.sect.data.push(format!("  {l_name}: .quad {value}\n")),
-      "FNC" => return Ok(l_name),
-      _ => return Err(format!("Internal Error: Unrecognized name: {name}").into()),
+      GVar::Int => self.data.push(format!("  .L{seed:x}: .quad {value}\n")),
+      GVar::Fnc => (),
     }
-    Ok(name_fmt)
+    self.global_seed = add(self.global_seed, 1)?;
+    Ok(Name { var: Global, seed })
   }
   /// Gets variable.
   pub(crate) fn get_var(&self, var_name: &str, pos: &Position) -> ErrOR<Json> {
@@ -158,32 +155,18 @@ impl Jsonpiler {
     err!(self, pos, "Undefined variables: `{var_name}`")
   }
   /// Registers a function in the function table.
-  pub(crate) fn register(&mut self, name: &str, flag: (bool, bool), j_func: JFunc) {
-    self.builtin.insert(name.into(), Builtin { skip_eval: flag.0, scoped: flag.1, func: j_func });
-  }
-  /// Converts `Json` (`StringVar` or `String`) to `StringVar`, otherwise return `Err`
-  pub(crate) fn string2var(
-    &mut self, json: JsonWithPos, ordinal: usize, func_name: &str,
-  ) -> ErrOR<String> {
-    if let Json::String(st) = json.value {
-      match st {
-        Bind::Lit(l_str) => self.get_global("STR", &l_str),
-        Bind::Var(var) => Ok(var),
-        Bind::Local(local) | Bind::Tmp(local) => Ok(fmt_local("qword", local)),
-      }
-    } else {
-      self.typ_err(ordinal, func_name, "String", &json)?;
-      Ok(String::new())
-    }
+  pub(crate) fn register(&mut self, name: &str, flag: (bool, bool), func: JFunc) {
+    self.builtin.insert(name.into(), Builtin { skip_eval: flag.0, scoped: flag.1, func });
   }
 }
 /// Call function and return value.
 fn call_func(pos: Position, af: &AsmFunc, info: &mut FuncInfo) -> ErrOR<JsonWithPos> {
-  info.body.push(format!("  call {}\n", af.name));
+  const MOV: &str = "  mov ";
+  info.body.push(format!("  call .L{:x}\n", af.name));
   if let Json::Int(_) = *af.ret {
-    let offset = info.get_local(Align::U64)?;
-    info.body.push(format!("  mov {}, rax\n", fmt_local("qword", offset)));
-    Ok(JsonWithPos { value: Json::Int(Bind::Tmp(offset)), pos })
+    let name = info.get_local(8)?;
+    info.body.push(format!("{MOV}qword{name}, rax\n"));
+    Ok(JsonWithPos { value: Json::Int(Bind::Var(Name { var: Tmp, seed: name.seed })), pos })
   } else {
     Ok(JsonWithPos { value: Json::Null, pos })
   }