+++ src/json.rs
@@ -1,57 +1,54 @@
 //! Implementation of the `Json`
-use super::{Bind, Json, JsonWithPos};
+use super::{
+  AsmBool, Bind, Json, JsonWithPos, Name,
+  Var::{self, Global, Local, Tmp},
+};
 use core::{
   fmt::{self, Write as _},
   mem::take,
 };
 impl Json {
-  /// Determines if it is a literal.
-  pub fn is_literal(&self) -> bool {
-    match self {
-      Json::LBool(_) | Json::Null => true,
-      Json::VBool(..) | Json::Function(_) => false,
-      Json::Object(object) => matches!(object, Bind::Lit(_)),
-      Json::Float(float) => matches!(float, Bind::Lit(_)),
-      Json::Int(int) => matches!(int, Bind::Lit(_)),
-      Json::String(string) => matches!(string, Bind::Lit(_)),
-      Json::Array(array) => matches!(array, Bind::Lit(_)),
-    }
-  }
   /// Determines if it is a temporary value.
   pub fn tmp(&self) -> Option<(usize, usize)> {
     match self {
       Json::LBool(_) | Json::Null | Json::VBool(_) | Json::Function(_) => None,
-      Json::Object(object) => match object {
-        Bind::Tmp(local) => Some((*local, 8)),
-        _ => None,
-      },
-      Json::Float(float) => match float {
-        Bind::Tmp(local) => Some((*local, 8)),
-        _ => None,
-      },
-      Json::Int(int) => match int {
-        Bind::Tmp(local) => Some((*local, 8)),
-        _ => None,
-      },
-      Json::String(string) => match string {
-        Bind::Tmp(local) => Some((*local, 8)),
-        _ => None,
-      },
-      Json::Array(array) => match array {
-        Bind::Tmp(local) => Some((*local, 8)),
-        _ => None,
-      },
+      Json::Object(bind) => bind_match(bind, 8),
+      Json::Float(bind) => bind_match(bind, 8),
+      Json::Int(bind) => bind_match(bind, 8),
+      Json::String(bind) => bind_match(bind, 8),
+      Json::Array(bind) => bind_match(bind, 8),
     }
   }
   /// Converts temporary value to Local variable.
   pub fn tmp_to_local(&mut self) -> Self {
-    match self {
-      Json::Object(Bind::Tmp(local)) => Json::Object(Bind::Local(*local)),
-      Json::Array(Bind::Tmp(local)) => Json::Array(Bind::Local(*local)),
-      Json::Float(Bind::Tmp(local)) => Json::Float(Bind::Local(*local)),
-      Json::String(Bind::Tmp(local)) => Json::String(Bind::Local(*local)),
-      Json::Int(Bind::Tmp(local)) => Json::Int(Bind::Local(*local)),
-      _ => take(self),
+    match *self {
+      Json::Object(Bind::Var(Name { var: Tmp, seed })) => {
+        Json::Object(Bind::Var(Name { var: Local, seed }))
+      }
+      Json::Array(Bind::Var(Name { var: Tmp, seed })) => {
+        Json::Array(Bind::Var(Name { var: Local, seed }))
+      }
+      Json::Float(Bind::Var(Name { var: Tmp, seed })) => {
+        Json::Float(Bind::Var(Name { var: Local, seed }))
+      }
+      Json::String(Bind::Var(Name { var: Tmp, seed })) => {
+        Json::String(Bind::Var(Name { var: Local, seed }))
+      }
+      Json::Int(Bind::Var(Name { var: Tmp, seed })) => {
+        Json::Int(Bind::Var(Name { var: Local, seed }))
+      }
+      Json::VBool(AsmBool { name: Name { var: Tmp, seed }, bit }) => {
+        Json::VBool(AsmBool { name: Name { var: Local, seed }, bit })
+      }
+      Json::String(_)
+      | Json::Int(_)
+      | Json::Object(_)
+      | Json::Float(_)
+      | Json::Array(_)
+      | Json::LBool(_)
+      | Json::VBool(_)
+      | Json::Function(_)
+      | Json::Null => take(self),
     }
   }
   /// Generate type name.
@@ -68,6 +65,25 @@ impl Json {
       Json::Array(arr) => bind_name(arr, "Array"),
     }
   }
+  /// Determines if it is a global variable.
+  pub fn var(&self) -> Option<Var> {
+    match self {
+      Json::Function(_) => Some(Global),
+      Json::VBool(AsmBool { name, .. })
+      | Json::Object(Bind::Var(name))
+      | Json::Float(Bind::Var(name))
+      | Json::Int(Bind::Var(name))
+      | Json::String(Bind::Var(name))
+      | Json::Array(Bind::Var(name)) => Some(name.var),
+      Json::String(_)
+      | Json::Int(_)
+      | Json::Object(_)
+      | Json::Float(_)
+      | Json::Array(_)
+      | Json::LBool(_)
+      | Json::Null => None,
+    }
+  }
 }
 impl fmt::Display for Json {
   /// Formats the `Json` object as a compact string without indentation.
@@ -81,25 +97,21 @@ impl fmt::Display for Json {
           iter_write(array, f)?;
           f.write_str("]")
         }
-        Bind::Var(var) => write!(f, "VArray(\"{var}\")"),
-        Bind::Local(local) | Bind::Tmp(local) => write!(f, "VArray(\"{local}\")"),
+        Bind::Var(name) => write!(f, "VArray(qword{name})"),
       },
       Json::LBool(bo) => write!(f, "{bo}"),
-      Json::VBool(vb) => write!(f, "VBool(\"{}\"-{})", vb.name, vb.bit),
+      Json::VBool(asm_bool) => write!(f, "VBool(qword{}-{})", asm_bool.name, asm_bool.bit),
       Json::Int(int) => match int {
         Bind::Lit(l_int) => write!(f, "{l_int}"),
-        Bind::Var(var) => write!(f, "VInt(\"{var}\")"),
-        Bind::Local(local) | Bind::Tmp(local) => write!(f, "VInt(\"{local}\")"),
+        Bind::Var(name) => write!(f, "VInt(qword{name})"),
       },
       Json::Float(float) => match float {
-        Bind::Lit(flt) => write!(f, "{flt}"),
-        Bind::Var(var) => write!(f, "VFloat(\"{var}\")"),
-        Bind::Local(local) | Bind::Tmp(local) => write!(f, "VFloat(\"{local}\")"),
+        Bind::Lit(l_float) => write!(f, "{l_float}"),
+        Bind::Var(name) => write!(f, "VFloat(qword{name})"),
       },
       Json::String(st) => match st {
-        Bind::Lit(lstr) => f.write_str(&escape_string(lstr)?),
-        Bind::Var(var) => write!(f, "VString(\"{var}\")"),
-        Bind::Local(local) | Bind::Tmp(local) => write!(f, "VString(\"{local}\")"),
+        Bind::Lit(l_st) => f.write_str(&escape_string(l_st)?),
+        Bind::Var(name) => write!(f, "VString(qword{name})"),
       },
       Json::Function(fu) => {
         write!(f, "{}(", fu.name)?;
@@ -119,10 +131,16 @@ impl fmt::Display for Json {
           }
           f.write_str("}")
         }
-        Bind::Var(var) => write!(f, "VObject({var})"),
-        Bind::Local(local) | Bind::Tmp(local) => write!(f, "VObject(\"{local}\")"),
+        Bind::Var(name) => write!(f, "VObject(qword{name})"),
       },
     }
+  }
+}
+/// Pattern match of the `Bind`.
+fn bind_match<T>(bind: &Bind<T>, byte: usize) -> Option<(usize, usize)> {
+  match bind {
+    Bind::Var(name) => Some((name.seed, byte)),
+    Bind::Lit(_) => None,
   }
 }
 /// gets name of the `Bind`.